minor fixes for invalid audio data patch by (Wolfram Gloger: wmglo, dent med uni-muenchen de)
Originally committed as revision 4524 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/resample2.c
@@ -139,7 +139,7 @@ AVResampleContext *av_resample_init(int out_rate, int in_rate, int filter_size, 
     c->phase_mask= phase_count-1;
     c->linear= linear;
 
-    c->filter_length= FFMAX(ceil(filter_size/factor), 1);
+    c->filter_length= FFMAX((int)ceil(filter_size/factor), 1);
     c->filter_bank= av_mallocz(c->filter_length*(phase_count+1)*sizeof(FELEM));
     av_build_filter(c->filter_bank, factor, c->filter_length, phase_count, 1<<FILTER_SHIFT, 1);
     memcpy(&c->filter_bank[c->filter_length*phase_count+1], c->filter_bank, (c->filter_length-1)*sizeof(FELEM));
